idek if this is valid syntax Lol
the magic of template metaprogramming am i right

this is not done, not to self: maybe restructure to make the indices part of the callconv class, since instantiating it a lot is too much copy paste, along with being very hard on the compiler

make another filter with the tuple in the lambda to grab the args that are correct. also, maybe making a RawArgs param pack could be useful as well? also very hard on the compiler to calculate so many times